+++ src/components/layout/landing-page/hero.tsx
